+++ commands/src/main/java/net/apartium/cocoabeans/commands/lexer/SimpleArgumentParserToken.java
@@ -3,6 +3,8 @@ package net.apartium.cocoabeans.commands.lexer;
 import net.apartium.cocoabeans.commands.RegisterArgumentParser;
 import net.apartium.cocoabeans.commands.parsers.ArgumentParser;
 import org.jetbrains.annotations.ApiStatus;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Map;
 import java.util.Objects;
@@ -18,6 +20,8 @@ import java.util.regex.Pattern;
 public class SimpleArgumentParserToken extends ArgumentParserToken {
 
     private static final Pattern PARAMETER_NAME_REGEX = Pattern.compile("^[a-zA-Z0-9_\\-]+:");
+
+    private static final Logger logger = LoggerFactory.getLogger(SimpleArgumentParserToken.class);
 
     private String parserKeyword;
     private String parameterName;
@@ -75,9 +79,26 @@ public class SimpleArgumentParserToken extends ArgumentParserToken {
      */
     @Override
     public RegisterArgumentParser<?> getParser(Map<String, ArgumentParser<?>> parsers) {
+        return getParser(parsers, null);
+    }
+
+    /**
+     * Gets the argument parser from the given parsers map
+     * @param parsers parser map that represents all argument parsers we have
+     * @param fallback fallback parser when couldn't find parser in map
+     * @return the argument parser or fallback
+     */
+    @ApiStatus.AvailableSince("0.0.41")
+    @Override
+    public RegisterArgumentParser<?> getParser(Map<String, ArgumentParser<?>> parsers, ArgumentParser<?> fallback) {
         ArgumentParser<?> argumentParser = parsers.get(parserKeyword);
-        if (argumentParser == null)
-            throw new IllegalArgumentException("Parser not found: " + parserKeyword);
+        if (argumentParser == null) {
+            if (fallback == null)
+                throw new IllegalArgumentException("Parser not found: " + parserKeyword);
+
+            logger.warn("Parser not found for: {} & fallback to: {}", parserKeyword, fallback.getClass().getSimpleName());
+            return new RegisterArgumentParser<>(fallback, optionalNotMatch, isOptional, parameterName);
+        }
 
         return new RegisterArgumentParser<>(argumentParser, optionalNotMatch, isOptional, parameterName);
     }